+++ CHANGES.md
@@ -7,6 +7,7 @@
   * Fixed Tooltip not using custom locale
   * Added option to export current custom theme as JSON string
   * Added option to clone built-in theme into custom colors for easy tweaking
+  * Added option to swap colors in custom theme editor
 
 * v1.3.0 (2020-12-18)
   * Added update availability checker